Run_airflow_in_docker_MSSQL_Postgres/requirements.txt
# For postgresql
apache-airflow-providers-postgres
psycopg2

# For MSSQL
apache-airflow-providers-odbc
apache-airflow-providers-microsoft-mssql
apache-airflow-providers-microsoft-mssql[odbc]
pyodbc

pandas
sqlalchemy